refactor/#44: comments swagger deco 파일명 변경
