+++ src/comments/comments.controller.ts
@@ -12,14 +12,14 @@ import {
 import { CommentService } from "./comments.service";
 import { CommentContentDto } from "./dtos/comment-dtos/comment-content.dto";
 import { ApiTags } from "@nestjs/swagger";
-import { ApiCreateComment } from "./swagger-decorators/comment-swagger/create-comment-decorator";
-import { ApiFindComments } from "./swagger-decorators/comment-swagger/find-comments-decorator";
-import { ApiUpdateComment } from "./swagger-decorators/comment-swagger/update-comment-decorator";
-import { ApiDeleteComment } from "./swagger-decorators/comment-swagger/delete-comments-decorator";
-import { ApiFindRelies } from "./swagger-decorators/reply-swagger/find-reply-decorate";
-import { ApiDeleteReply } from "./swagger-decorators/reply-swagger/delete-reply-decorate";
-import { ApiUpdateReply } from "./swagger-decorators/reply-swagger/update-reply-decorate";
-import { ApiCreateReply } from "./swagger-decorators/reply-swagger/create-reply-decorater";
+import { ApiCreateComment } from "./swagger-decorators/comment-swagger/create-comment.decorator";
+import { ApiFindComments } from "./swagger-decorators/comment-swagger/find-comments.decorator";
+import { ApiUpdateComment } from "./swagger-decorators/comment-swagger/update-comment.decorator";
+import { ApiDeleteComment } from "./swagger-decorators/comment-swagger/delete-comments.decorator";
+import { ApiFindRelies } from "./swagger-decorators/reply-swagger/find-reply.decorator";
+import { ApiDeleteReply } from "./swagger-decorators/reply-swagger/delete-reply.decorator";
+import { ApiUpdateReply } from "./swagger-decorators/reply-swagger/update-reply.decorator";
+import { ApiCreateReply } from "./swagger-decorators/reply-swagger/create-reply.decorator";
 import { ParsePositiveIntPipe } from "src/common/pipes/parse-positive-int.pipe";
 import { PaginationDto } from "src/common/dtos/pagination.dto";
